+++ source/Ashes/VkRenderer/Src/Miscellaneous/VkDeviceMemory.cpp
@@ -41,8 +41,8 @@ namespace vk_renderer
 		m_device.vkFreeMemory( m_device, m_memory, nullptr );
 	}
 
-	uint8_t * DeviceMemory::lock( uint32_t offset
-		, uint32_t size
+	uint8_t * DeviceMemory::lock( uint64_t offset
+		, uint64_t size
 		, ashes::MemoryMapFlags flags )const
 	{
 		uint8_t * pointer{ nullptr };
@@ -56,8 +56,8 @@ namespace vk_renderer
 		return pointer;
 	}
 
-	void DeviceMemory::flush( uint32_t offset
-		, uint32_t size )const
+	void DeviceMemory::flush( uint64_t offset
+		, uint64_t size )const
 	{
 		VkMappedMemoryRange mappedRange
 		{
@@ -72,8 +72,8 @@ namespace vk_renderer
 		checkError( res, "DeviceMemory range flush" );
 	}
 
-	void DeviceMemory::invalidate( uint32_t offset
-		, uint32_t size )const
+	void DeviceMemory::invalidate( uint64_t offset
+		, uint64_t size )const
 	{
 		VkMappedMemoryRange mappedRange
 		{
